# [CONSVC-1603]: add feature flags as tags to metrics