--- a/merino/metrics.py
+++ b/merino/metrics.py
@@ -2,10 +2,9 @@
 
 import logging
 from functools import cache
-from typing import Any, Callable, Final, Mapping, TypeVar, Union
+from typing import Any, Callable, Concatenate, Final, Mapping, ParamSpec, TypeVar, Union
 
-from aiodogstatsd import Client as StatsDClient
-from aiodogstatsd.client import DatagramProtocol
+import aiodogstatsd
 from wrapt import decorator
 
 from merino.config import settings
@@ -14,12 +13,9 @@ from merino.featureflags import FeatureFlags
 logger = logging.getLogger(__name__)
 
 # Type definitions for tags in aiodogstatsd metrics
-MTagKey = str
-MTagValue = Union[float, int, str]
-MTags = Mapping[MTagKey, MTagValue]
-
-# TypeVar for the `ClientMeta` meta class
-_M = TypeVar("_M", bound="ClientMeta")
+MetricTagKey = str
+MetricTagValue = Union[float, int, str]
+MetricTags = Mapping[MetricTagKey, MetricTagValue]
 
 # Type definitions for the `calls` attribute on the `Client` class
 MetricCall = dict[str, str | tuple | dict]
@@ -40,7 +36,7 @@ SUPPORTED_METHODS: Final[list[str]] = [
 FLAGS_PREFIX: Final[str] = "feature_flag"
 
 
-def feature_flags_as_tags(feature_flags: FeatureFlags) -> MTags:
+def feature_flags_as_tags(feature_flags: FeatureFlags) -> MetricTags:
     """Return a representation of feature flags decisions."""
     return {
         f"{FLAGS_PREFIX}.{name}": int(decision)
@@ -51,7 +47,7 @@ def feature_flags_as_tags(feature_flags: FeatureFlags) -> MTags:
 @decorator
 def add_feature_flags(
     wrapped_method: Callable, instance: "Client", args: tuple, kwargs: dict
-):
+) -> Any:
     """Add feature flag decisions as tags when recording metrics."""
     # Tags added manually to the metrics client call
     tags = kwargs.pop("tags", {})
@@ -66,35 +62,52 @@ def add_feature_flags(
     return wrapped_method(*args, **kwargs)
 
 
+# TypeVar for the `ClientMeta` meta class
+M = TypeVar("M", bound="ClientMeta")
+
+# TypeVar for the `Client` class
+C = TypeVar("C", bound="Client")
+
+# ParamSpec for the client_method
+P = ParamSpec("P")
+
+# TypeVar for the return type of the StatsD client method
+R = TypeVar("R")
+
+
 class ClientMeta(type):
     """Metaclass that decorates Client methods with add_feature_flags."""
 
     def __new__(
-        cls: type[_M],
+        mcls: type[M],
         name: str,
         bases: tuple[type, ...],
         namespace: dict[str, Any],
-        **kwargs: Any,
-    ) -> _M:
+        **kwargs: dict[str, Any],
+    ) -> M:
         """Create a new class with decorated methods."""
 
-        def method_proxy(method_name: str) -> Callable:
+        def method_proxy(
+            method_name: str,
+        ) -> Callable[Concatenate[C, P], R]:
             """Return a method that proxies to correct method of the StatsD client."""
 
-            def client_method(instance: "Client", *args, **kwargs) -> Any:
+            def client_method(
+                instance: C, *method_args: P.args, **method_kwargs: P.kwargs
+            ) -> R:
                 """Look up the correct method of the StatsD client call it."""
                 # Keep track of all calls made to the metrics client
                 call: MetricCall = {
                     "method_name": method_name,
-                    "args": args,
-                    "kwargs": kwargs,
+                    "args": method_args,
+                    "kwargs": method_kwargs,
                 }
                 instance.calls.append(call)
 
                 # Look up the method on the StatsD client on the instance
-                method: Callable = getattr(instance.statsd_client, method_name)
+                method: Callable[..., R] = getattr(instance.statsd_client, method_name)
 
-                return method(*args, **kwargs)
+                return method(*method_args, **method_kwargs)
 
             return client_method
 
@@ -102,18 +115,18 @@ class ClientMeta(type):
         for method_name in SUPPORTED_METHODS:
             namespace[method_name] = add_feature_flags(method_proxy(method_name))
 
-        return super().__new__(cls, name, bases, namespace, **kwargs)
+        return super().__new__(mcls, name, bases, namespace, **kwargs)
 
 
 class Client(metaclass=ClientMeta):
     """Proxy for a StatsD client that adds tags for feature flags."""
 
-    statsd_client: StatsDClient
+    statsd_client: aiodogstatsd.Client
     feature_flags: FeatureFlags
     calls: MetricCalls
 
     def __init__(
-        self, statsd_client: StatsDClient, feature_flags: FeatureFlags
+        self, statsd_client: aiodogstatsd.Client, feature_flags: FeatureFlags
     ) -> None:
         """Initialize the client instance."""
         self.statsd_client = statsd_client
@@ -129,9 +142,9 @@ class Client(metaclass=ClientMeta):
 
 
 @cache
-def get_metrics_client() -> StatsDClient:
+def get_metrics_client() -> aiodogstatsd.Client:
     """Instantiate and memoize the StatsD client."""
-    return StatsDClient(
+    return aiodogstatsd.Client(
         host=settings.metrics.host,
         port=settings.metrics.port,
         namespace="merino",
@@ -147,7 +160,7 @@ async def configure_metrics() -> None:
     await client.connect()
 
 
-class _LocalDatagramLogger(DatagramProtocol):
+class _LocalDatagramLogger(aiodogstatsd.client.DatagramProtocol):
     """
     This class can be used to override the default DatagramProtocol.
     Instead of writing bytes to a socket, it logs them.