--- a/merino/middleware/__init__.py
+++ b/merino/middleware/__init__.py
@@ -8,3 +8,5 @@ class ScopeKey(str, Enum):
 
     GEOLOCATION = "merino_geolocation"
     USER_AGENT = "merino_user_agent"
+    FEATURE_FLAGS: str = "merino_feature_flags"
+    METRICS_CLIENT: str = "merino_metrics_client"