--- a/pyproject.toml
+++ b/pyproject.toml
@@ -65,6 +65,7 @@ aiodogstatsd = "^0.16.0"
 ua-parser = "^0.16.1"
 geoip2 = "^4.6.0"
 rich = "^12.5.1"
+wrapt = "^1.14.1"
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.1.2"