## Description

This pull request integrates feature flags with StatsD metrics. 📈

Feature flag decisions are now automatically recorded and added as tags to emitted metrics.

## Issue(s)

[CONSVC-1603](https://mozilla-hub.atlassian.net/browse/CONSVC-1603)