@ncloudioj 2022-11-02
@hackebrot Sorry for the delayed response.

I've skimmed over the PR, and it looks good. I will take another closer look just in a bit.

Also, I've run some preliminary benchmarks locally, and have seen some performance improvements 🚀 so far:

Running 100VU for 60 seconds.

The main branch: 2,531 RPS, 52.18ms latency (P95)
This branch: 2,858 RPS, 47.2ms

Will do it with a few more realistic scenarios to verify the results.