@ncloudioj 2022-11-03
Okay, more benchmark results:

Tl;dr: local benchmarks show that this branch outperforms (10% relative uplift w.r.t throughput & latency) the main branch in both testing scenarios (w/ & w/o feature flags).

## Without feature flags
### Main branch
- RPS: 2,531
- Latency (P95): 52.18ms

### This branch
- RPS: 2,858
- Latency (P95): 47.20ms

## With three feature flags
All results below are using the following feature flag configs:
```
[development.flags.feature-a]
enabled = 1
scheme = 'session'

[development.flags.feature-b]
enabled = 0.5
scheme = 'session'

[development.flags.feature-c]
enabled = 0.5
scheme = 'session'
```

### Main branch with 3 feature flags
 - RPS: 2,504
- Latency (P95): 58.00ms

### This branch with 3 feature flags
 - RPS: 2,814
- Latency (P95): 53.00ms